// pronti-frontend-main/front/src/app/modulos/trabajador/servicios/dto/trabajador.response-dto.ts
import {AbstractResponseDto} from "../../../../abstract/dto/abstract-response.dto";
import {ModalidadContratoEnum} from '../../../../enums/modalidad-contrato';
import {UsuarioResponseDto} from '../../../usuario/servicios/dto/usuario.response-dto';
import {SupervisorResponseDto} from '../../../supervisor/servicios/dto/supervisor.response-dto';
import {GeneroEnum} from '../../../../enums/genero.enum';
import {EstadoFamiliarEnum} from '../../../../enums/estado-familiar.enum';
import {SiNoEnum} from '../../../../enums/si-no.enum';
import {AgenciaResponseDto} from "../../../agencia/servicios/dto/agencia.response-dto";
import { InformacionFinancieraResponseDto } from '../../../informacion-financiera/servicios/dto/informacion-financiera.response-dto';
import { CargaFamiliarResponseDto } from '../../../carga-familiar/servicios/dto/carga-familiar.response-dto';
import { HistorialLaboralResponseDto } from '../../../historial-laboral/servicios/dto/historial-laboral.response-dto';

export interface TrabajadorResponseDto extends AbstractResponseDto {
  // nombre?: string;
  modalidadContrato?: ModalidadContratoEnum;
  idUsuario?: UsuarioResponseDto;
  nombreCompleto?: string;
  idSupervisor?: SupervisorResponseDto
  estadoCivil?: string;
  genero?: GeneroEnum;
  grupoSanguineo?: string;
  nivelEstudios?: string;
  profesion?: string;
  direccionDomiciliaria?: string;
  estadoFamiliar?: EstadoFamiliarEnum;
  numeroAfiliacion?: string;
  discapacidad?: SiNoEnum;
  tipoDiscapacidad?: string;
  sueldo?: number;
  bonificacion?: number;
  movilizacionEspecial?: number;
  componenteSalarialUnifi?: number;
  retencionesJudiciales?: number;
  polizaPersonal?: number;
  aporteIess?: SiNoEnum;
  aporteIessAsumido?: SiNoEnum;
  decimosAnio?: SiNoEnum;
  impuestoRenta?: SiNoEnum;
  fondoReservaIess?: SiNoEnum;
  utilidades?: SiNoEnum;
  pagoFondosReservaMes?: SiNoEnum;
  impuestoRentaPatron?: SiNoEnum;
  vacaciones?: SiNoEnum;
  beneficios?: SiNoEnum;
  pagoDecTercerCuartoMes?: SiNoEnum;
  cuentaContableNombre?: string;
  acumuladoAportePersonalIess?: number;
  acumuladoDecimoTercero?: number;
  acumuladoFondosReserva?: number;
  acumuladoVacaciones?: number;
  acumuladoImpuestoRenta?: number;
  utilidadCargas?: number;
  tiempoParcial?: SiNoEnum;
  factorParcial?: number;
  pasante?: SiNoEnum;
  reingreso?: SiNoEnum;
  fechaReingreso?: string;
  provDecimoTercero?: number;
  provDecimoCuarto?: number;
  provFondosReserva?: number;
  provVacaciones?: number;
  provAportePatronal?: number;
  fechaIngreso?: string;
  decimoCuarto?: number;
  idAgencia?: AgenciaResponseDto;
  informacionFinancieraCollection?: InformacionFinancieraResponseDto[];
  cargaFamiliarCollection?: CargaFamiliarResponseDto[];
  historialLaboralCollection?: HistorialLaboralResponseDto[];

}