// pronti-frontend-main/front/src/app/modulos/rol-pago/servicios/dto/rol-pago.response-dto.ts
import {AbstractResponseDto} from "../../../../abstract/dto/abstract-response.dto";
import {PeriodoLaboralResponseDto} from "../../../periodo-laboral/servicios/dto/periodo-laboral.response-dto";
import {HistorialLaboralResponseDto} from "../../../historial-laboral/servicios/dto/historial-laboral.response-dto";
import { HistoricoRolResponseDto } from '../../../historico-rol/servicios/dto/historico-rol.response-dto';

export interface RolPagoResponseDto extends AbstractResponseDto {
  // nombre?: string;
  horas25?: number;
  horas50?: number;
  horas100?: number;
  retencionesJudiciales?: number;
  polizaPersonal?: number;
  aporteIess?: number;
  impuestoRenta?: number;
  descuentosPorAtrasos?: number;
  anticipos?: number;
  multas?: number;
  otrosDescuentos?: number;
  celularConsumo?: number;
  permisoHoras?: number;
  permisoDias?: number;
  descuentosPorFaltas?: number;
  leySolidaria?: number;
  prestamosEmpresa?: number;
  peluqueria?: number;
  optica?: number;
  pagoCuotaVehiculo?: number;
  sueldo?: number;
  bonificacion?: number;
  movilizacionEspecial?: number;
  componenteSalarialUnif?: number;
  diasFalta?: number;
  diasMaternidad?: number;
  diasEnfermedad?: number;
  diasVacacion?: number;
  horasAtrado?: number;
  descuentoMaternidad?: number;
  decuentoEnfermedad?: number;
  descuentoVacacion?: number;
  provDecimoTercero?: number;
  provDecimoCuarto?: number;
  provFondosReserva?: number;
  provVacaciones?: number;
  provAportePatronal?: number;
  vacacion1?: number;
  vacacion2?: number;
  pagoVacacion?: number;
  pagoFondoReservaMes?: number;
  pagoDecimoTerceroCuartoMes?: number;
  acumuladoAportePersonalIess?: number;
  acumuladoDecimoTercero?: number;
  acumuladoFondosReserva?: number;
  acumuladoVacaciones?: number;
  acumuladoImpuestoRenta?: number;
  totalIngresos?: number;
  totalEgresos?: number;
  totalAPagar?: number;
  otrosIngresos?: number;
  prestamosIess?: number;
  totalHorasExtra?: number;
  fechaDesde?: string;
  fechaHasta?: string;
  comision?: number;
  pagoDecimoCuartoMes?: number;
  prestamoQuirografario?: number;
  prestamoHipotecario?: number;
  idPeriodoLaboral?: PeriodoLaboralResponseDto;
  idHistorialLaboral?: HistorialLaboralResponseDto;
  idHistoricoRol?: any;
  utilidad?:number;
}